requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytest-mock==3.10.0
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.12.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
